+++ lib/src/state_management/state_management.dart
@@ -4,6 +4,7 @@ export 'computed_shard.dart';
 export 'debounce_throttle.dart';
 export 'equality.dart';
 export 'future_shard.dart';
+export 'history.dart';
 export 'persistent_shard.dart';
 export 'shard.dart';
 export 'shard_observer.dart';